// explore/explore-ui/projects/explore-ui-core/src/favorite/enums/favorite-display.enum.ts
// Enum containing entry for each class that extends AbstractFavoriteConfig.  Used to look-up the UI display name of the favorite.
export enum FavoriteDisplayEnum {
    ADHOC_PORT_GROUP = 'ADHOC_PORT_GROUP',
    ADHOC_PORT = 'ADHOC_PORT',
    BATCH_REPORT = 'BATCH_REPORT',
    BREAKDOWN = 'BREAKDOWN',
    COLUMN_SET = 'COLUMN_SET',
    COMP_RULE = 'COMP_RULE',
    COMPOSITION_SETTING = 'COMPOSITION_SETTING',
    CUSTOM_CALC = 'CUSTOM_CALC',
    CUSTOM_SECTOR = 'CUSTOM_SECTOR',
    FILTER = 'FILTER',
    LOOK_THROUGH_SETTING = 'LOOK_THROUGH_SETTING',
    LOOK_THROUGH_RULE = 'LOOK_THROUGH_RULE',
    OPTIMIZATION_SETTINGS = 'OPTIMIZATION_SETTINGS',
    PORTFOLIO = 'PORTFOLIO',
    PORT_WITH_POSITIONS = 'PORT_WITH_POSITIONS',
    REPORT = 'REPORT',
    RISK_PARITY_SETTINGS = 'RISK_PARITY_SETTINGS',
    RULE_BASED_PORT = 'RULE_BASED_PORT',
    SCHEDULED_BATCH = 'SCHEDULED_BATCH',
    WHAT_IF_PORTFOLIO = 'WHAT_IF_PORTFOLIO',
    WORKSPACE = 'WORKSPACE'
}